+++ src/common/js/configure.js
@@ -2,7 +2,7 @@
  * pwix:roles/src/common/js/configure.js
  */
 
-import merge from 'merge';
+import _ from 'lodash';
 
 pwixRoles._defaults = {
     roles: {},
@@ -10,17 +10,25 @@ pwixRoles._defaults = {
     verbosity: PR_VERBOSE_NONE
 };
 
+/**
+ * @summary Get/set the package configuration
+ *  Should be called *in same terms* both in the client and the server
+ * @locus Anywhere
+ * @param {Object} o configuration options
+ * @returns {Object} the package configuration
+ */
 pwixRoles.configure = function( o ){
-    pwixRoles._conf = merge.recursive( true, pwixRoles._defaults, o );
-
-    // be verbose if asked for
-    if( pwixRoles._conf.verbosity & PR_VERBOSE_CONFIGURE ){
-        console.debug( 'pwix:roles configure() with', o, 'building', pwixRoles._conf );
+    if( o && _.isObject( o )){
+        _.merge( pwixRoles._conf, pwixRoles._defaults, o );
+        // be verbose if asked for
+        if( pwixRoles._conf.verbosity & PR_VERBOSE_CONFIGURE ){
+            console.debug( 'pwix:roles configure() with', o, 'building', pwixRoles._conf );
+        }
+        if( Meteor.isClient ){
+            pwixRoles._client.currentRecompute( Meteor.userId());
+        }
     }
-
-    if( Meteor.isClient ){
-        pwixRoles._client.currentRecompute( Meteor.userId());
-    }
+    return pwixRoles._conf;
 }
 
-pwixRoles._conf = merge.recursive( true, pwixRoles._defaults );
+_.merge.recursive( pwixRoles._conf, pwixRoles._defaults );