+++ src/server/js/check_npms.js
@@ -9,7 +9,7 @@ if( false ){
 
 checkNpmVersions({
     'deep-equal': '^2.2.0',
-    'merge': '^2.1.1',
+    'lodash': '^4.17.0',
     'uuid': '^9.0.0'
 },
     'pwix:roles'